python_script/requirements.txt
paho-mqtt
django
plotly
